pyproject.toml
[project]
name = "bnb-hack"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.0.1",
    "cryptography>=41.0.7",
    "eth-account>=0.11.0",
    "eth-hash>=0.5.2",
    "flask>=2.3.3",
    "flask-cors>=4.0.0",
    "flask-jwt-extended>=4.5.3",
    "python-dotenv>=1.0.0",
    "qrcode[pil]>=7.4.2",
    "requests>=2.31.0",
    "setuptools>=80.9.0",
    "web3>=6.11.3",
    "werkzeug>=2.3.7",
]
